Dev

See merge request nft/nft-indexer-api!26

+++ src/token.rs
@@ -19,7 +19,7 @@ pub struct Token {
     pub vendor: Option<String>,
     #[serde(rename = "logoURI")]
     pub logo_uri: String,
-    pub decimals: usize,
+    pub decimals: u32,
     pub verified: bool,
 }
